remove MarkAggregateOrderRule
